Re-sync-up.

+++ lib/Migrator/Migrator.cpp
@@ -26,22 +26,26 @@
 using namespace swift;
 using namespace swift::migrator;
 
-bool migrator::updateCodeAndEmitRemap(CompilerInstance *Instance,
-                                      const CompilerInvocation &Invocation) {
+bool migrator::updateCodeAndEmitRemapIfNeeded(
+    CompilerInstance *Instance, const CompilerInvocation &Invocation) {
+  if (!Invocation.getMigratorOptions().shouldRunMigrator())
+    return false;
+
   // Delete the remap file, in case someone is re-running the Migrator. If the
   // file fails to compile and we don't get a chance to overwrite it, the old
   // changes may get picked up.
   llvm::sys::fs::remove(Invocation.getMigratorOptions().EmitRemapFilePath);
 
   Migrator M { Instance, Invocation }; // Provide inputs and configuration
+  auto EffectiveVersion = Invocation.getLangOptions().EffectiveLanguageVersion;
+  auto CurrentVersion = version::Version::getCurrentLanguageVersion();
 
   // Phase 1: Pre Fix-it passes
   // These uses the initial frontend invocation to apply any obvious fix-its
   // to see if we can get an error-free AST to get to Phase 2.
   std::unique_ptr<swift::CompilerInstance> PreFixItInstance;
   if (Instance->getASTContext().hadError()) {
-    PreFixItInstance = M.repeatFixitMigrations(2,
-      Invocation.getLangOptions().EffectiveLanguageVersion);
+    PreFixItInstance = M.repeatFixitMigrations(2, EffectiveVersion);
 
     // If we still couldn't fix all of the errors, give up.
     if (PreFixItInstance == nullptr ||
@@ -53,7 +57,8 @@ bool migrator::updateCodeAndEmitRemap(CompilerInstance *Instance,
   }
 
   // Phase 2: Syntactic Transformations
-  if (Invocation.getLangOptions().EffectiveLanguageVersion[0] < 4) {
+  // Don't run these passes if we're already in newest Swift version.
+  if (EffectiveVersion != CurrentVersion) {
     auto FailedSyntacticPasses = M.performSyntacticPasses();
     if (FailedSyntacticPasses) {
       return true;
@@ -69,7 +74,7 @@ bool migrator::updateCodeAndEmitRemap(CompilerInstance *Instance,
 
   if (M.getMigratorOptions().EnableMigratorFixits) {
     M.repeatFixitMigrations(Migrator::MaxCompilerFixitPassIterations,
-                            {4, 0, 0});
+                            CurrentVersion);
   }
 
   // OK, we have a final resulting text. Now we compare against the input
@@ -127,20 +132,22 @@ Migrator::performAFixItMigration(version::Version SwiftLanguageVersion) {
     LLVMArgs.erase(aarch64_use_tbi);
   }
 
-  // SE-0160: When migrating, always use the Swift 3 @objc inference rules,
-  // which drives warnings with the "@objc" Fix-Its.
-  Invocation.getLangOptions().EnableSwift3ObjCInference = true;
+  if (StartInvocation.getLangOptions().EffectiveLanguageVersion.isVersion3()) {
+    // SE-0160: When migrating, always use the Swift 3 @objc inference rules,
+    // which drives warnings with the "@objc" Fix-Its.
+    Invocation.getLangOptions().EnableSwift3ObjCInference = true;
 
-  // The default behavior of the migrator, referred to as "minimal" migration
-  // in SE-0160, only adds @objc Fix-Its to those cases where the Objective-C
-  // entry point is explicitly used somewhere in the source code. The user
-  // may also select a workflow that adds @objc for every declaration that
-  // would infer @objc under the Swift 3 rules but would no longer infer
-  // @objc in Swift 4.
-  Invocation.getLangOptions().WarnSwift3ObjCInference =
-    getMigratorOptions().KeepObjcVisibility
-      ? Swift3ObjCInferenceWarnings::Complete
-      : Swift3ObjCInferenceWarnings::Minimal;
+    // The default behavior of the migrator, referred to as "minimal" migration
+    // in SE-0160, only adds @objc Fix-Its to those cases where the Objective-C
+    // entry point is explicitly used somewhere in the source code. The user
+    // may also select a workflow that adds @objc for every declaration that
+    // would infer @objc under the Swift 3 rules but would no longer infer
+    // @objc in Swift 4.
+    Invocation.getLangOptions().WarnSwift3ObjCInference =
+      getMigratorOptions().KeepObjcVisibility
+        ? Swift3ObjCInferenceWarnings::Complete
+        : Swift3ObjCInferenceWarnings::Minimal;
+  }
 
   const auto &OrigFrontendOpts = StartInvocation.getFrontendOptions();
